FossilOrigin-Name: b9fad4490c437a2c49cf86cc1c068433d6d2866b

+++ ext/fts3/fts3_expr.c
@@ -502,7 +502,7 @@ static int getNextNode(
       return rc;
     }else if( *zInput==')' ){
       pParse->nNest--;
-      *pnConsumed = (zInput - z) + 1;
+      *pnConsumed = (int)((zInput - z) + 1);
       *ppExpr = 0;
       return SQLITE_DONE;
     }
